Cleaned up custom item menu item into new component
